main: create the output directory if needed
The examples from the website use the "images" directory, but it
doesn't exist in a newly-cloned repository and it is not obvious that
it should be manually created beforehand.

To make it easier for the user, create the output directory if needed.

Note that the C++17 'std::filesystem::create_directory' is
intentionally not used in order to avoid requiring C++17 only for this
change.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>